🦭 Reuse background job controls in workspace
