+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - **后台任务面板跟随任务状态更顺手**：触发后台任务后会自动展开右侧后台任务面板；运行中的实时输出会智能跟随滚到底部，并在用户手动翻看历史输出时保留位置、提供回到底部入口；每个后台任务条目现在也可单独展开 / 收起，避免长输出把整块面板撑满。
+- **工作台后台任务可直接展开操作**：工作台里的后台任务区块现在复用后台任务面板同一套任务行能力，可直接展开查看实时输出 / 结果预览、取消仍在运行的任务，并与独立后台任务面板共享展开状态；工作台仍保留紧凑展示和「查看全部」入口。
 - **文件编辑失败不再误显行数变更**：`edit` / `apply_patch` 的 `+N/-M` 参数预估只在工具运行中展示；工具完成后必须使用后端真实 diff metadata，避免失败或无操作编辑看起来像已经改动文件。
 - **工作台 Git 变更可直接查看 diff**：点击工作台「环境」里的「变更」行现在会打开右侧 diff 面板，按文件纵向展示当前会话工作目录内的 Git 改动，并支持逐文件收起 / 展开；server 模式下该接口也收紧到会话工作目录范围，符号链接按 Git 链接文本展示，不会跟随读取工作区外文件。
 - **查看子智能体对话不再切走主对话**：除左侧对话列表外，消息气泡、团队面板、后台任务面板与工作台里的子智能体对话入口现在统一打开实时弹层查看，并复用主对话消息流能力；在消息对话页查看子智能体时不会再把当前主 Agent 会话切走。
